+++ dbutils/pooled_db.py
@@ -114,6 +114,13 @@ connection object stays alive as long as you are using it, like that:
     cur.close()  # or del cur
     db.close()  # or del db
 
+You can also use context managers for simpler code:
+
+    with pool.connection() as db:
+        with db.cursor as cur:
+            cur.execute(...)
+            res = cur.fetchone()
+
 Note that you need to explicitly start transactions by calling the
 begin() method.  This ensures that the connection will not be shared
 with other threads, that the transparent reopening will be suspended
@@ -440,6 +447,14 @@ class PooledDedicatedDBConnection:
         except:  # builtin Exceptions might not exist any more
             pass
 
+    def __enter__(self):
+        """Enter a runtime context for the connection."""
+        return self
+
+    def __exit__(self, *exc):
+        """Exit a runtime context for the connection."""
+        self.close()
+
 
 class SharedDBConnection:
     """Auxiliary class for shared connections."""
@@ -526,3 +541,11 @@ class PooledSharedDBConnection:
             self.close()
         except:  # builtin Exceptions might not exist any more
             pass
+
+    def __enter__(self):
+        """Enter a runtime context for the connection."""
+        return self
+
+    def __exit__(self, *exc):
+        """Exit a runtime context for the connection."""
+        self.close()