+++ docs/changelog.rst
@@ -1,6 +1,16 @@
 ﻿Changelog for DBUtils
 +++++++++++++++++++++
 
+2.0.2
+=====
+
+DBUtils 2.0.2 was released on ...
+
+Changes:
+
+* Allow using context managers for pooled connections.
+
+
 2.0.1
 =====
 
@@ -8,7 +18,7 @@ DBUtils 2.0.1 was released on April 8, 2021.
 
 Changes:
 
-* Avoid "name Exception is not defined" when exiting
+* Avoid "name Exception is not defined" when exiting.
 
 2.0
 ===